Fixed water shader, fixed UI cutoff textures, adjusted visuals for dynamic 3D tiles, ambient occlusion, etc.

+++ core/src/main/java/dev/michey/expo/render/ui/InteractableItemSlot.java
@@ -58,7 +58,7 @@ public class InteractableItemSlot extends InteractableUIElement {
                 float _x = this.x + (parent.slotW - mapping.uiRender[0].useWidth * parent.uiScale) * 0.5f;
                 float _y = this.y + (parent.slotH - mapping.uiRender[0].useHeight * parent.uiScale) * 0.5f;
 
-                r.hudBatch.draw(ir.useTextureRegion, _x + parent.uiScale + parent.uiScale * ir.offsetX, _y - parent.uiScale * 2 + parent.uiScale * ir.offsetY,
+                r.hudBatch.draw(ir.useTextureRegion, (int) (_x + parent.uiScale + parent.uiScale * ir.offsetX), (int) (_y - parent.uiScale * 2 + parent.uiScale * ir.offsetY),
                         ir.useTextureRegion.getRegionWidth() * parent.uiScale, ir.useTextureRegion.getRegionHeight() * parent.uiScale);
             }
 
@@ -67,7 +67,7 @@ public class InteractableItemSlot extends InteractableUIElement {
                 float _x = this.x + (parent.slotW - mapping.uiRender[0].useWidth * parent.uiScale) * 0.5f;
                 float _y = this.y + (parent.slotH - mapping.uiRender[0].useHeight * parent.uiScale) * 0.5f;
 
-                r.hudBatch.draw(ir.useTextureRegion, _x + parent.uiScale * ir.offsetX, _y + parent.uiScale * ir.offsetY,
+                r.hudBatch.draw(ir.useTextureRegion, (int) (_x + parent.uiScale * ir.offsetX), (int) (_y + parent.uiScale * ir.offsetY),
                         ir.useTextureRegion.getRegionWidth() * parent.uiScale, ir.useTextureRegion.getRegionHeight() * parent.uiScale);
             }
